movie-explorer-backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
django-cors-headers==4.6.0
